+++ src/SymbolTable.ts
@@ -1,3 +1,5 @@
+import * as AST from './AST';
+
 export type EntryType = 'name' | 'function';
 
 export interface PrimarySymbolTableEntry {
@@ -6,13 +8,13 @@ export interface PrimarySymbolTableEntry {
 
 export interface ValueSymbolTableEntry extends PrimarySymbolTableEntry {
     type: 'name';
-    expr?: any;
+    expr?: AST.NodeExpr;
 }
 
 export interface FunctionSymbolTableEntry extends PrimarySymbolTableEntry {
     type: 'function';
-    args?: any[];
-    expr?: any;
+    args?: AST.NodeIdentifier[];
+    expr?: AST.NodeExpr;
 }
 
 export type SymbolTableEntry = ValueSymbolTableEntry | FunctionSymbolTableEntry;